move dotproduct tests from sourmash_lib/__init__.py to test__minhash.py
